requirements.txt
numpy==1.26.2
flake8==6.1.0
isort==5.13.0
black==23.11.0
environs~=10.0.0
python-dotenv~=1.0.0
polars~=0.15
